# Conflicts:
#	modules/core/src/main/java/org/apache/ignite/internal/managers/communication/GridIoMessageFactory.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/distributed/dht/preloader/IgniteDhtPartitionsToReloadMap.java

+++ modules/core/src/test/java/org/apache/ignite/internal/util/distributed/MessagesPluginProvider.java
@@ -34,8 +34,8 @@ import static org.junit.Assert.assertTrue;
 public class MessagesPluginProvider extends AbstractTestPluginProvider {
     /** */
     private static final MessageFactoryProvider FACTORY_PROVIDER = f -> {
-        f.register((short)10_000, TestIntegerMessage::new, new TestIntegerMessageSerializer());
-        f.register((short)10_001, TestUuidMessage::new, new TestUuidMessageSerializer());
+        f.register(10_000, TestIntegerMessage::new, new TestIntegerMessageSerializer());
+        f.register(10_001, TestUuidMessage::new, new TestUuidMessageSerializer());
     };
 
     /** {@inheritDoc} */
